+++ components/projects.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const ProjectsComponent = () => {
+    return (
+        <section className='mb-28 max-w-[50rem] text-center sm:mb-0'>
+            Bismillah
+        </section>
+    )
+}
+
+export default ProjectsComponent;